requirements.txt
boto3>=1.34
ffmpeg-python>=0.2
openai>=1.0
python-dotenv>=1.0
fastapi[all]>=0.111
